+++ packages/ice-npm-utils/src/__tests__/index.ts
@@ -16,9 +16,9 @@ import {
   getAndExtractTarball,
 } from '../index';
 
-const defaultRegistry = 'https://registry.npm.taobao.org';
-// 可能会返回不同的源
-const tbRegisties = [defaultRegistry, 'https://registry.npmmirror.com'];
+const defaultRegistry = 'https://registry.npmmirror.com';
+// 预留 cnpm 返回多个源的可能性
+const registies = [defaultRegistry];
 
 jest.setTimeout(10 * 1000);
 
@@ -88,7 +88,7 @@ test('getNpmInfo success', () => {
 
 test('getNpmInfo 404 error case', () => {
   return getNpmInfo('not-exis-npm-error').catch((err) => {
-    expect(err.response.status).toBe(404);
+    expect(err.statusCode).toBe(404);
   });
 });
 
@@ -114,7 +114,7 @@ test('getNpmClient', () => {
   expect(getNpmClient('@alixxx/ice-test')).toBe(defaultData);
 });
 
-test.only('checkAliInternal', () => {
+test('checkAliInternal', () => {
   return checkAliInternal().then((internal) => {
     console.log('checkAliInternal', internal);
     expect(internal).toBeBoolean();
@@ -126,15 +126,8 @@ test('getNpmTarball', () => {
   return getNpmTarball('ice-npm-utils', '1.0.0').then((tarball) => {
     console.log('getNpmTarball ice-npm-utils', tarball);
     expect(
-      tbRegisties
-        .some(registry => tarball === `${registry}/ice-npm-utils/download/ice-npm-utils-1.0.0.tgz`)
+      registies.some(registry => tarball === `${registry}/ice-npm-utils/-/ice-npm-utils-1.0.0.tgz`)
     ).toBeTruthy();
-  });
-});
-
-test('getNpmTarball 404', () => {
-  return getNpmTarball('not-exis-npm-error').catch((err) => {
-    expect(err.response.status).toBe(404);
   });
 });
 
@@ -142,22 +135,17 @@ test('getNpmTarball should get latest version', () => {
   return getNpmTarball('http').then((tarball) => {
     console.log('getNpmTarball http', tarball);
     expect(
-      tbRegisties
-        .some(registry => tarball === `${registry}/http/download/http-0.0.1-security.tgz`)
+      registies.some(registry => tarball === `${registry}/http/-/http-0.0.1-security.tgz`)
     ).toBeTruthy();
   });
 });
 
 test('getAndExtractTarball', () => {
   const tempDir = path.resolve(tmpdir(), 'ice_npm_utils_tarball');
-  let percent;
-  return getAndExtractTarball(tempDir, `${defaultRegistry}/ice-npm-utils/download/ice-npm-utils-1.0.0.tgz`, (state) => {
-    percent = state.percent;
-  })
+  return getAndExtractTarball(tempDir, `${defaultRegistry}/ice-npm-utils/-/ice-npm-utils-1.0.0.tgz`)
     .then((files) => {
       rimraf.sync(tempDir);
       expect(files.length > 0).toBe(true);
-      expect(percent).toBe(1);
     })
     .catch(() => {
       rimraf.sync(tempDir);