+++ packages/template-component-demo/src/index.jsx
@@ -4,25 +4,37 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import * as queryString from 'query-string';
 import { Playground } from './Playground';
 import { BuildLayout, Layout } from './Layout';
+import { ReactDoc } from './ReactDoc';
 
-const App = ({ demoData, readmeData }) => {
-  const demos= readmeData ? [].concat(readmeData, demoData) : demoData;
+const App = ({ demoData, readmeData, reactDocs }) => {
+  const demos = readmeData ? [].concat(readmeData, demoData) : demoData;
   const items = (demos || []).sort((a, b) => a.order - b.order);
-  
+  const hasReactDocs = !!(reactDocs && reactDocs.length);
+
   return (
-    <BuildLayout demos={items}>
+    <BuildLayout demos={items} reactDocs={reactDocs}>
       {items.map((data) => {
         const Comp = data.component;
         return (
           <div key={data.filename}>
             <a name={`container_${data.filename}`} />
             {data.title && data.filename !== 'readme' && <h3>{data.title}</h3>}
-            <Playground data={data}>
-              {typeof Comp === 'function' && <Comp />}
-            </Playground>
+            <Playground data={data}>{typeof Comp === 'function' && <Comp />}</Playground>
           </div>
         );
       })}
+      {
+        hasReactDocs && (
+          <div>
+            <h2>API</h2>
+            <a name="container_api" />
+            {reactDocs.map((data) => (
+              <ReactDoc key={data.filename} data={data} />
+            ))}
+          </div>
+        )
+      }
+
     </BuildLayout>
   );
 };
@@ -32,7 +44,7 @@ const Demo = ({ location, demoData, readmeData }) => {
   const demos = demoData.find((item) => item[0] && item[0].demoKey === demo);
   return (
     <Layout demos={[].concat(readmeData, demoData)} matchedFilename={demo}>
-      { demos ? <App demoData={demos} /> : <Playground data={readmeData} /> }
+      {demos ? <App demoData={demos} /> : <Playground data={readmeData} />}
     </Layout>
   );
 };